+++ frontend/src/pages/MainPage/MainPage.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import React from 'react';
 import { useParams } from 'react-router-dom';
 import MembersProfile from '../../components/MembersProfile/MembersProfile';
-import MainFeatureMenuContainer from '../../components/Main/MainFeatureMenuContainer';
+import MainFeatureMenuContainer from '../../components/Main/MainFeatureMenuContainer/MainFeatureMenuContainer';
 
 function MainPage() {
   const { groupCode } = useParams() as { groupCode: string };
